# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student Portal API",
        "description": "This is the API documentation for the Student Portal backend.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/academic-years": {
            "get": {
                "tags": [
                    "Academic Year"
                ],
                "summary": "List all academic years",
                "description": "Get all academic years",
                "operationId": "ea3748d03aaf9f0996c64306142aab38",
                "responses": {
                    "200": {
                        "description": "List of academic years",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AcademicYear"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academic-years": {
            "post": {
                "tags": [
                    "Academic Year"
                ],
                "summary": "Create a new academic year",
                "description": "Store a new academic year",
                "operationId": "9b60bfda0678f44df4b8454fdb33c1bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2024/2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Academic year created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicYear"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academic-years/{id}": {
            "get": {
                "tags": [
                    "Academic Year"
                ],
                "summary": "Get academic year by ID",
                "description": "Get details of a specific academic year",
                "operationId": "9bfeb72edf70711264ff0e46984ac3b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the academic year",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic year details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicYear"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Academic Year"
                ],
                "summary": "Update an academic year",
                "description": "Update an existing academic year",
                "operationId": "69c8bbb7e26cf4a4ce9fa6227f0a736a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the academic year to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "2025/2026"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Academic year updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcademicYear"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Academic Year"
                ],
                "summary": "Delete an academic year",
                "description": "Delete an academic year",
                "operationId": "b68045f9d242e62de2e7b04964c7040b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the academic year to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Academic year deleted"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "description": "Authenticate user and return a token",
                "operationId": "1e24877eaace300cc29c3b260060d729",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "your-jwt-token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AcademicYear": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "2024/2025"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-05T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-05T12:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Academic Year",
            "description": "API endpoints for managing academic years"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}